refactor: unify login flow and cleanup
