+++ src/lib/supabase.ts
@@ -5,3 +5,5 @@ export const createSupabaseBrowser = () =>
     process.env.NEXT_PUBLIC_SUPABASE_URL || 'http://localhost',
     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY || 'anon'
   );
+
+export const supabase = createSupabaseBrowser();
